Starting to polish everything up.
Added interface to the interactable items so i dont need to cast.
Added Enum to know what type of action i have.
Added InteractableItem and InteractableDress. Item is to be removed next.
Added in the contructors of the Actors the set up of Tags so we don't have to do it manually every time we add an object to the scene.
Added an interaction function to the player.